requirements.txt
Django==2.2.14
django-taggit==2.0.0
isort==4.3.21
Pillow==7.2.0
pycodestyle==2.6.0
pydocstyle==5.0.2
pylint==2.4.4
pytest==5.4.3
Markdown==3.3.6
psycopg2==2.8.6
django-braces==1.15.0
django-embed-video==1.4.1
python-memcached==1.59
django-memcache-status==2.2
djangorestframework==3.13.1
